% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB_to_Fast.R
\name{cBtofast}
\alias{cBtofast}
\title{Extract data for efficient analysis from cB}
\usage{
cBtofast(
  cB_raw,
  c_list,
  type_list,
  mut_list,
  rep_list,
  tl,
  nreps,
  keep_input = c(0.2, 50)
)
}
\arguments{
\item{c_list}{vector of names of control (no s4U fed) samples}

\item{type_list}{vector with 1 entry per sample; 0 = no s4U, 1 = s4U fed}

\item{mut_list}{vector with 1 entry per sample; 1 = WT, > 1 = different experimental conditions (e.g., KO of protein X)}

\item{rep_list}{vector with 1 entry per sample that indexes replicate; 1 = 1st replicate, 2 = 2nd replicate, etc.}

\item{tl}{single numerical value; s4U label time used in s4U fed samples}

\item{nreps}{single numerical value; number of replicates (assumes same number of replicates for each mut_list index)}

\item{keep_input}{two element vector; 1st element is highest mut rate accepted in control samples, 2nd element is read count cutoff}

\item{cB}{cB file generated from TL-seq pipeline}

\item{samp_list}{vector of names of all samples}
}
\value{
returns dataframe that can be passed to fast analysis
}
\description{
This function processes the cB into a form necessary for the parameter estimation function
that does not call Stan and is thus much more efficient and scalable.
}
