% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicSeqData.R
\name{DynamicSeqData}
\alias{DynamicSeqData}
\title{DynamicSeq Data object helper function for users}
\usage{
DynamicSeqData(cB, metadf)
}
\arguments{
\item{cB}{Dataframe with columns corresponding to feature ID, number of Ts, number of mutations, sample ID, and number of identical observations}

\item{metadata}{Dataframe detailing s4U label time and experimental ID of each sample}
}
\description{
This function creates an object of class DynamicSeqData
}
