% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fast_analysis.R
\name{fast_analysis}
\alias{fast_analysis}
\title{Efficiently analyze nucleotide recoding data}
\usage{
fast_analysis(
  df,
  pnew = NULL,
  pold = NULL,
  read_cut = 50,
  features_cut = 10,
  nbin = NULL,
  prior_weight = 2,
  MLE = TRUE,
  lower = -7,
  upper = 7,
  se_max = 2.5
)
}
\arguments{
\item{df}{Dataframe in form provided by cB_to_Fast}

\item{pnew}{Labeled read mutation rate; default of 0 means that model estimates rate from s4U fed data}

\item{pold}{Unlabeled read mutation rate; default of 0 means that model estimates rate from no-s4U fed data}

\item{read_cut}{Minimum number of reads for a given feature-sample combo to be used for mut rate estimates}

\item{features_cut}{Number of features to estimate sample specific mutation rate with}

\item{nbin}{Number of bins for mean-variance relationship estimation. If NULL, max of 10 or (number of logit(fn) estimates)/100 is used}

\item{prior_weight}{Determines extent to which logit(fn) variance is regularized to the mean-variance regression line}

\item{MLE}{Logical; if TRUE then replicate logit(fn) is estimated using maximum likelihood; if FALSE more conservative Bayesian hypothesis testing is used}

\item{lower}{Lower bound for MLE with L-BFGS-B algorithm}

\item{upper}{Upper bound for MLE with L-BFGS-B algorithm}
}
\value{
list with dataframe of replicate specific estimates as well as dataframe of pooled estimates
}
\description{
\code{fast_analysis} analyzes nucleotide recoding data using either maximum likelihood estimation with the L-BFGS-B algorithm
implemented by \code{stats::optim} or a more conservative Bayesian hypothesis testing strategy. Output includes fraction new
estimates for individual estimates as well as fraction news and effect sizes (L2FC(kdeg)s and changes in logit(fraction new))
averaged across replicate data. Averaging takes into account uncertainties estimated using the Fisher Information and estimates
are reguarlized using analytical results of fully Bayesian models. The result is that fraction news are shrunk towards population means
and that uncertainties are shrunk towards a mean-variance trend estimated as part of the analysis.
}
