% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{FnPCA}
\alias{FnPCA}
\title{Creating PCA plots with logit(fn) estimates}
\usage{
FnPCA(obj, ...)
}
\arguments{
\item{obj}{Object of class FastFit or HMCFit outputted by respective analysis functions}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This function creates a 2-component PCA plot with logit(fn) estimates.
}
