% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakRData.R
\name{new_bakRData}
\alias{new_bakRData}
\title{bakRData object constructor for internal use}
\usage{
new_bakRData(cB, metadf)
}
\arguments{
\item{cB}{Dataframe with columns corresponding to feature ID, number of Ts, number of mutations, sample ID, and number of identical observations}

\item{metadf}{Dataframe detailing s4U label time and experimental ID of each sample}
}
\description{
This function efficiently creates an object of class bakRData
without performing rigorous checks
}
