% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakRData.R
\name{new_bakRFnData}
\alias{new_bakRFnData}
\title{bakRFnData object constructor for internal use}
\usage{
new_bakRFnData(fns, metadf)
}
\arguments{
\item{fns}{Dataframe with columns corresponding to sample names (sample), feature IDs (XF),
fraction new estimates (fn), and number of sequencing reads (nreads)}

\item{metadf}{Dataframe detailing s4U label time and experimental ID of each sample}
}
\description{
This function efficiently creates an object of class bakRFnData
without performing rigorous checks
}
