% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSSHeat.R
\name{NSSHeat}
\alias{NSSHeat}
\title{Extract data for efficient analysis from cB}
\usage{
NSSHeat(
  bakRFit,
  DE_df,
  bakRModel = c("MLE", "Hybrid", "MCMC"),
  DE_cutoff = 0.05,
  bakR_cutoff = 0.05,
  Exp_ID = 2,
  lid = 4
)
}
\arguments{
\item{bakRFit}{bakRFit object}

\item{DE_df}{dataframe of required format with differential expression analysis results. See
Further-Analyses vignette for details on what this dataframe should look like}

\item{bakRModel}{Model fit from which bakR implementation should be used? Options are MLE, Hybrid,
or MCMC}

\item{DE_cutoff}{padj cutoff for calling a gene differentially expressed}

\item{bakR_cutoff}{padj cutoff for calling a fraction new significantly changed}

\item{Exp_ID}{Exp_ID of experimental sample whose comparison to the reference sample you want to use.
Only one reference vs. experimental sample comparison can be used at a time}

\item{lid}{Maximum absolute value for standardized score present in output. This is for improving
aesthetics of any heatmap generated with the output.}
}
\value{
returns dataframe that can be passed to pheatmap::pheatmap()
}
\description{
This function processes the cB into a form necessary for the parameter estimation function
that does not call Stan and is thus much more efficient and scalable.
}
