% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{plotVolcano.FastFit}
\alias{plotVolcano.FastFit}
\title{S3 method for class 'FastFit'}
\usage{
\method{plotVolcano}{FastFit}(obj, FDR = 0.05, Exps = NULL, Exp_shape = FALSE, ...)
}
\arguments{
\item{FDR}{False discovery rate to control at for significance assessment}

\item{Exps}{Vector of Experimental IDs to include in plot; must only contain elements within 2:(# of experimental IDs)}

\item{Exp_shape}{Logical indicating whether to use Expeirmental ID as factor determining point shape in volcano plot}
}
\description{
S3 method for class 'FastFit'
}
