% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{plotVolcano}
\alias{plotVolcano}
\title{S3 generic for creating Volcano Plots}
\usage{
plotVolcano(obj, ...)
}
\arguments{
\item{obj}{Object of class FastFit or HMCFit outputted by respective analysis functions}

\item{...}{Further arguments pass to or from other methods}
}
\description{
This function performs method dispatch to create a L2FC(kdeg) volcano plot.
Methods are available for objects of class HMCFit (generated by TL_stan() function) and
objects of class FastFit (generated by fast_analysis() function). Plots are colored according
to statistical significance and sign of L2FC(kdeg).
}
