% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC_checks}
\alias{QC_checks}
\title{Check data quality and make suggestions to user about what analyses to run.}
\usage{
QC_checks(obj)
}
\arguments{
\item{obj}{bakRFit object}
}
\description{
\code{QC_checks} takes as input a bakRFit object and uses the Fast_Fit object to assess
data quality and make suggestions about which implementation to run next. QC_checks
takes into account the mutation rates in all samples, the fraction new distributions, the reproducibility
of fraction new estimates, and the read lengths. It then outputs a number of
diagnostic plots that might alert users to problems in their data. It also
outputs messages informing users what implementation is best used next.
}
