% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{sim_DynamicSeqData}
\alias{sim_DynamicSeqData}
\title{Simulating nucleotide recoding data}
\usage{
sim_DynamicSeqData(
  ngene,
  num_conds = 2L,
  nreps = 3L,
  eff_sd = 0.75,
  eff_mean = 0,
  fn_sd = 1,
  tl = 60,
  p_new = 0.05,
  p_old = 0.001,
  read_lengths = 200L,
  p_do = 0,
  noise_deg_a = -0.3,
  noise_deg_b = -1.5,
  noise_synth = 0.1,
  sd_rep = 0.05,
  low_L2FC_ks = -1,
  high_L2FC_ks = 1,
  num_kd_DE = c(0L, as.integer(rep(round(as.integer(ngene)/2), times =
    as.integer(num_conds) - 1))),
  num_ks_DE = rep(0L, times = as.integer(num_conds)),
  scale_factor = 100,
  sim_read_counts = TRUE,
  a1 = 5,
  a0 = 0.01,
  nreads = 50L
)
}
\arguments{
\item{ngene}{Number of genes to simulate data for}

\item{num_conds}{Number of experimental conditions (including the reference condition) to simulate}

\item{nreps}{Number of replicates to simulate}

\item{eff_sd}{Effect size; more specifically, the standard deviation of the normal distribution from which non-zero
changes in logit(fraction new) are pulled from.}

\item{eff_mean}{Effect size mean; mean of normal distribution from which non-zero changes in logit(fraction new) are pulled from.
Note, setting this to 0 does not mean that some of the significant effect sizes will be 0, as any exact integer is impossible
to draw from a continuous random number generator. Setting this to 0 just means that there is symmetric stabilization and destabilization}

\item{tl}{s4U label feed time}

\item{p_new}{s4U induced mutation rate. Can be a vector of length num_conds}

\item{p_old}{background mutation rate}

\item{read_lengths}{Total read length for each sequencing read (e.g., 200 means PE100 reads)}

\item{p_do}{Rate at which s4U containing reads are lost due to dropout; must be between 0 and 1}

\item{noise_deg_a}{Slope of trend relating log10(standardized read counts) to log(replicate variability)}

\item{noise_deg_b}{Intercept of trend relating log10(standardized read counts) to log(replicate variability)}

\item{noise_synth}{Homoskedastic variability of L2FC(ksyn)}

\item{sd_rep}{Variance of lognormal distribution from which replicate variability is drawn}

\item{low_L2FC_ks}{Most negative L2FC(ksyn) that can be simulated}

\item{high_L2FC_ks}{Most positive L2FC(ksyn) that can be simulated}

\item{num_kd_DE}{Vector where each element represents the number of genes that show a significant change in stability relative
to the reference. 1st entry must be 0 by definition (since relative to the reference the reference sample is unchanged)}

\item{num_ks_DE}{Same as num_kd_DE but for significant changes in synthesis rates.}

\item{scale_factor}{Factor relating RNA concentration (in arbitrary units) to average number of read counts}

\item{sim_read_counts}{Logical; if TRUE, read counts are simulated as coming from a heterodisperse negative binomial distribution}

\item{a1}{Heterodispersion 1/reads dependence parameter}

\item{a0}{High read depth limit of negative binomial dispersion parameter}

\item{nread}{Number of reads simulated if sim_read_counts is FALSE}
}
\value{
A list containing a simulated \code{DynamicSeqData} object as well as a list of simulated kinetic parameters of interest.
The contents of the latter list are:
\itemize{
\item L2FC_kd_mean; The true L2FC(kdeg) for each feature in each experimental condition. The ith column corresponds to the L2FC(kdeg) when comparing
the ith condition to the reference condition (defined as the 1st condition) so the 1st column is always all 0s
\item fn_mean; True fraction new for each feature in each experimental condition.
\item fn_true; The fraction new used for each replicate of each feature in each experimental condition. These are generated using a heteroskedastic
logit-normal distribution, with variance depending on the abundance of the feature (greater RNA concentration = less variability)
\item effect_mean; The true change in logit(fraction new) for each feature in each experimental condition. The ith column corresponds to the
change in logit(fraction new) when comparing the ith condition to the reference condition (defined as the 1st condition), so the 1st column is
always all 0s
\item L2FC_ks_mean; The true L2FC(ksyn) for each feature in each experimental condition. The ith column corresponds to the L2FC(ksyn) when comparing
the ith condition to the reference condition (defined as the 1st condition) so the 1st column is always all 0s
\item RNA_conc; The average number of normalized read counts expected for each feature in each sample.
}
}
\description{
\code{sim_DynamicSeqData} simulates a \code{DynamicSeqData} object. It's output also includes the simulated
values of all kinetic parameters of interest. Only the number of genes (\code{ngene}) has to be set by the
user, but an extensive list of additional parameters can be adjusted.
}
\details{
\code{sim_DynamicSeqData} simulates a \code{DynamicSeqData} object using an extensive list of
adjustable parameters. Average RNA kinetic parameters are drawn from biologically inspired
distributions. Replicate variability is simulated by drawing each replicate and feature's
fraction labeled (aka fraction new) from a Logit-Normal distribution with a heteroskedastic
variance term with average magnitude given by a read count vs. variance relationship, and each
replicate and feature's ksyn from a homoskedastic lognormal distribution. Read counts
can either be set to the same value for all simulated features or can be simulated according to
a heterodisperse negative binomial distribution.

The number of Us in each sequencing read is drawn from a binomial distribution with number of trials
equal to the read length and probability of each nucleotide being a U of 0.25. Each read is assigned to the
labeled or unlabeled population according to a Bernoulli distribution with p = fraction new. The number of
mutations in each read are then drawn from one of two binomial distributions; if the read is assigned to the
labeled population, the number of mutations are drawn from a binomial distribution with number of trials equal
to the number of Us and probability of mutation = \code{p_new}; if the read is assigned to the unlabeled population,
the number of mutations is instead drawn from a binomial distribution with the same number of trials but with the probability
of mutation = \code{p_old}.

Simulated read counts should be treated as if they are spike-in and RPKM normalized, so the same scale factor can be applied
to each sample when comparing the sequencing reads (like if you are performing differential expression analysis).

Function to simulate a DynamicSeqData object according to a heteroskedastic beta-binomial model
of the fraction new.
}
