% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicSeqData.R
\name{validate_bakRData}
\alias{validate_bakRData}
\title{bakR Data object validator}
\usage{
validate_bakRData(obj)
}
\arguments{
\item{obj}{An object of class bakRData}
}
\description{
This functions ensures that input for bakRData object construction is valid
}
