% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakRData.R
\name{validate_bakRFnData}
\alias{validate_bakRFnData}
\title{bakRFnData object validator}
\usage{
validate_bakRFnData(obj)
}
\arguments{
\item{obj}{An object of class bakRFnData}
}
\description{
This functions ensures that input for bakRFnData object construction is valid
}
