% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicSeqData.R
\name{validate_DynamicSeqData}
\alias{validate_DynamicSeqData}
\title{DynamicSeq Data object validator}
\usage{
validate_DynamicSeqData(obj)
}
\arguments{
\item{obj}{An object of class DynamicseqData}
}
\description{
This functions ensures that input for DynamicSeqData object construction is valid
}
