% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{VisualizeDropout}
\alias{VisualizeDropout}
\title{Visualize dropout}
\usage{
VisualizeDropout(obj, Exp_ID = 1, Replicate = 1)
}
\arguments{
\item{obj}{bakRFit or bakRFnFit object}

\item{Exp_ID}{Experimental condition ID to visualize dropout for}

\item{Replicate}{Replicate ID to visualize dropout for}
}
\description{
\code{VisualizeDropout} fits dropout model with \code{QuantifyDropout},
reports the fit results, and then generates a ggplot object showing the
data used to infer the fit as well as the fitted nonlinear trend.
}
