% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBprocessing.R
\name{cBprocess}
\alias{cBprocess}
\title{Extract data for Stan analysis from cB}
\usage{
cBprocess(
  cB_raw,
  samp_list,
  type_list,
  mut_list,
  rep_list,
  tl,
  keep_input = c(0.2, 50),
  Stan = TRUE,
  Fast = FALSE,
  FOI = c(),
  concat = TRUE
)
}
\arguments{
\item{cB_raw}{cB file generated from TL-seq pipeline}

\item{samp_list}{vector of names of samples}

\item{type_list}{vector with 1 entry per sample; 0 = no s4U, 1 = s4U fed}

\item{mut_list}{vector with 1 entry per sample; 1 = reference, > 1 = different experimental conditions (e.g., KO of protein X)}

\item{rep_list}{vector with 1 entry per sample that indexes replicate; 1 = 1st replicate, 2 = 2nd replicate, etc.}

\item{tl}{single numerical value; s4U label time used in s4U fed samples}

\item{keep_input}{two element vector; 1st element is highest mut rate accepted in control samples, 2nd element is read count cutoff}

\item{Stan}{Boolean; if TRUE, then data_list that can be passed to Stan is curated}

\item{Fast}{Boolean; if TRUE, then dataframe that can be passed to fast_analysis() is curated}

\item{FOI}{Features of interest; character vector containing names of features to analyze}

\item{concat}{Boolean; If TRUE, FOI is concatenated with output of reliableFeatures}
}
\value{
returns list which can be passed to Stan analyses
}
\description{
This function obtains the data list necessary to analyze TL-seq data with Stan.
}
