% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fast_analysis.R
\name{cBtofast}
\alias{cBtofast}
\title{Extract data for efficient analysis from cB}
\usage{
cBtofast(
  cB_raw,
  samp_list,
  type_list,
  mut_list,
  rep_list,
  tl,
  keep_input = c(0.2, 50),
  FOI = c(),
  concat = TRUE
)
}
\arguments{
\item{cB_raw}{cB file generated from TL-seq pipeline}

\item{samp_list}{vector of names of control samples}

\item{type_list}{vector with 1 entry per sample; 0 = no s4U, 1 = s4U fed}

\item{mut_list}{vector with 1 entry per sample; 1 = reference, > 1 = different experimental conditions (e.g., KO of protein X)}

\item{rep_list}{vector with 1 entry per sample that indexes replicate; 1 = 1st replicate, 2 = 2nd replicate, etc.}

\item{tl}{single numerical value; s4U label time used in s4U fed samples}

\item{keep_input}{two element vector; 1st element is highest mut rate accepted in control samples, 2nd element is read count cutoff}

\item{FOI}{Features of interest; character vector containing names of features to analyze}

\item{concat}{Boolean; If TRUE, FOI is concatenated with output of reliableFeatures}
}
\value{
returns dataframe that can be passed to fast analysis
}
\description{
This function processes the cB into a form necessary for the parameter estimation function
that does not call Stan and is thus much more efficient and scalable.
}
