% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB_to_Fast.R
\name{fast_analysis}
\alias{fast_analysis}
\title{This function efficiently analyzes data to provide fraction new estimates and uncertainties}
\usage{
fast_analysis(df, boot_iter = 50)
}
\arguments{
\item{df}{Dataframe in form provided by cB_to_Fast}

\item{boot_iter}{Number of times to resample for bootstrapping; default is 50}
}
\value{
list with dataframe of replicate specific estimates as well as dataframe of pooled estimates
}
\description{
This function efficiently analyzes data to provide fraction new estimates and uncertainties
}
