% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{FnPCA}
\alias{FnPCA}
\title{Creating PCA plots with logit(fn) estimates}
\usage{
FnPCA(obj, log_kdeg = FALSE)
}
\arguments{
\item{obj}{Object contained within output of \code{bakRFit}. So, either Fast_Fit (MLE implementation fit),
Stan_Fit (MCMC implementation fit), or Hybrid_Fit (Hybrid implementation fit)}

\item{log_kdeg}{Boolean; if TRUE, then log(kdeg) estimates used for PCA rather than logit(fn). Currently
only compatible with Fast_Fit}
}
\description{
This function creates a 2-component PCA plot using logit(fn) estimates.
}
