% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{FnPCA}
\alias{FnPCA}
\title{S3 generic for creating PCA with logit(fn)}
\usage{
FnPCA(obj, ...)
}
\arguments{
\item{obj}{Object of class FastFit or HMCFit outputted by respective analysis functions}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This function performs method dispatch to create a 2-component
PCA plot with logit(fn) estimates. Methods are available for
objects of class HMCFit (generated by TL_stan function) and
objects of class FastFit (generated by fast_analysis function)
}
