% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{Heatmap_kdeg}
\alias{Heatmap_kdeg}
\title{Creating a L2FC(kdeg) matrix that can be passed to heatmap functions}
\usage{
Heatmap_kdeg(obj, zscore = FALSE)
}
\arguments{
\item{obj}{Object of class FastFit or HMCFit outputted by respective analysis functions}

\item{zscore}{Logical; if TRUE, then each matrix entry is log-odds fold change in the fraction new (a.k.a the effect size) divided by
the uncertainty in the effect size}
}
\description{
\code{Heatmap_kdeg} creates a matrix where each column represents a pair of samples (reference and experimental) and each
row represents a feature. The entry in the ith row and jth column is the L2FC(kdeg) for feature i when comparing sample with
experimental ID j+1 to the reference sample
}
