% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{plotMA}
\alias{plotMA}
\title{Creating L2FC(kdeg) MA plot from fit objects}
\usage{
plotMA(obj, Avg_reads_natural, FDR = 0.05, Exps = NULL, Exp_shape = FALSE, ...)
}
\arguments{
\item{obj}{Object of class FastFit or HMCFit outputted by respective analysis functions}

\item{Avg_reads_natural}{Average reads across all replicates on natural linear scale; Obtained from \code{Fit$Data_Lists$Stan_data$Avg_reads_natural}}

\item{FDR}{False discovery rate to control at for significance assessment}

\item{Exps}{Vector of Experimental IDs to include in plot; must only contain elements within 2:(# of experimental IDs)}

\item{Exp_shape}{Logical indicating whether to use Expeirmental ID as factor determining point shape in volcano plot}
}
\description{
This function outputs a L2FC(kdeg) MA plot. Plots are colored according to statistical
significance and the sign of L2FC(kdeg)
}
