% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dropout_correction.R
\name{QuantifyDropout}
\alias{QuantifyDropout}
\title{Fit dropout model to quantify dropout frequency}
\usage{
QuantifyDropout(obj, keep_data = FALSE, ...)
}
\arguments{
\item{obj}{bakRFit object}

\item{keep_data}{Logical; if TRUE, will return list with two elements. First element
is the regular return (data frame with dropout quantitied), and the second element
will be the data frame that was used for fitting the dropout model. This is useful
if wanting to visualize the fit.}

\item{...}{Additoinal (optional) parameters to be passed to \code{stats::nls()}}
}
\description{
\code{QuantifyDropout} estimates the percentage of 4-thiouridine containing RNA
that was lost during library preparation (pdo).
}
