% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB_to_Stan.R
\name{reliableFeatures}
\alias{reliableFeatures}
\title{Identify features (e.g., transcripts) with high quality data}
\usage{
reliableFeatures(cB, c_list, high_p = 0.2, totcut = 1000)
}
\arguments{
\item{cB}{cB.rds file from pipeline}

\item{c_list}{vector of names of control (no s4U fed) samples}

\item{high_p}{highest mutation rate accepted in control samples}

\item{totcut}{readcount cutoff}
}
\value{
list of gene names that passed reliability filter
}
\description{
This function identifies all features (e.g., transcripts, exons, etc.) for which the mutation rate
is below a set threshold in the control (no s4U) sample and which have more reads than a set threshold
in at least one control sample
}
