% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBprocessing.R
\name{reliableFeatures}
\alias{reliableFeatures}
\title{Identify features (e.g., transcripts) with high quality data}
\usage{
reliableFeatures(obj, high_p = 0.2, totcut = 50, Ucut = 0.25, AvgU = 4)
}
\arguments{
\item{obj}{Object of class bakRData}

\item{high_p}{highest mutation rate accepted in control samples}

\item{totcut}{readcount cutoff}

\item{Ucut}{Must have a fraction of reads with 2 or less Us less than this cutoff in all samples}

\item{AvgU}{Must have an average number of Us greater than this}
}
\value{
vector of gene names that passed reliability filter
}
\description{
This function identifies all features (e.g., transcripts, exons, etc.) for which the mutation rate
is below a set threshold in the control (no s4U) sample and which have more reads than a set threshold
in all samples. If there is no -s4U sample, then only the read count cutoff is considered. Additional
filtering options are only relevant if working with short RNA-seq read data.
}
