% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat_inference.R
\name{StanFDR}
\alias{StanFDR}
\title{Assess statistical significance of fit object parameter}
\usage{
StanFDR(fit, pars, nconds, reps, FDR = 0.05, threshold = 0)
}
\arguments{
\item{fit}{fit object generated by TL_stan}

\item{pars}{name of parameter to perform statistical inference with; type string}

\item{nconds}{number of experimental conditions (nMT in data_list)}

\item{reps}{number of replicates}

\item{FDR}{level to control FDR at; default is 0.05}

\item{threshold}{null hypothesis is magnitude less than threshold; default is 0, the standard null hypothesis}
}
\value{
returns dataframe containing results of significance analysis
}
\description{
This function processes the cB into a form necessary for the parameter estimation function
that does not call Stan and is thus much more efficient and scalable.
}
