% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_stan.R
\name{TL_stan}
\alias{TL_stan}
\title{Bayesian hierarchical mixture model with Stan for TL-seq data analysis}
\usage{
TL_stan(data_list, ...)
}
\arguments{
\item{data_list}{list to pass to Stan of form given by cBtoStan}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian hierarchical mixture model with Stan for TL-seq data analysis
}
